Bug Fixed: Testing user existence in sign up

+++ base/api/views.py
@@ -49,13 +49,19 @@ class List_of_registered_user(APIView):
         return Response(serializer.data)
         
 class New_user_registration(APIView): 
-    def post(self, request,):
+    def post(self, request):
             email = request.data.get("email")
             password = request.data.get("password")
+            user_name = request.data.get("user_name")
+            user = NewUserRegistration.objects.filter(email=email)
+            if user.exists:
+                return Response({'message':'Email already exists'}, status=status.HTTP_400_BAD_REQUEST)
+            username = NewUserRegistration.objects.filter(user_name=user_name)
+            if username.exists:
+                return Response({'message':'Username already exists'}, status=status.HTTP_400_BAD_REQUEST)
             serializer = Verify_OTP_serializer(data=request.data)
             
             userOTP = OTP.objects.filter(email=email)
-            user = NewUserRegistration.objects.filter(email=email)
             
             if userOTP.exists() and not user.exists():
                 userOTP.delete()
